refactor: organize props & types
